% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_occurrence_count}
\alias{add_occurrence_count}
\title{Add Occurrence Count to Generated List}
\usage{
add_occurrence_count(
  data,
  track_by,
  count_col_name = "Occurrence",
  reset_by = NULL
)
}
\arguments{
\item{data}{Data frame with generated list.}

\item{track_by}{Column name for items to track.}

\item{count_col_name}{Name for the new occurrence count column.}

\item{reset_by}{Optional. Column name to reset counting (e.g., "Block").}
}
\value{
Data frame with added occurrence count column
}
\description{
Add item occurrence count to generated list
}
