% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyze_repetition_distribution}
\alias{analyze_repetition_distribution}
\title{Analyze Repetition Effects from Occurrence Data}
\usage{
analyze_repetition_distribution(data, occurrence_col = "Occurrence")
}
\arguments{
\item{data}{Data frame with occurrence count.}

\item{occurrence_col}{Name of occurrence count column.}
}
\description{
Analyze Repetition Effects from Occurrence Data
}
