% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_add_occurrence_count}
\alias{batch_add_occurrence_count}
\title{Batch Add Occurrence Count to Multiple Lists}
\usage{
batch_add_occurrence_count(
  results,
  track_by,
  count_col_name = "Occurrence",
  reset_by = NULL
)
}
\arguments{
\item{results}{List of results.}

\item{track_by}{Column name for items to track.}

\item{count_col_name}{Name for the occurrence count column.}

\item{reset_by}{Optional. Column to reset counting.}
}
\value{
Modified results list.
}
\description{
Batch Add Occurrence Count to Multiple Lists
}
