% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{flatten_constraints}
\alias{flatten_constraints}
\title{Flatten Nested Constraint Lists}
\usage{
flatten_constraints(constraints)
}
\arguments{
\item{constraints}{Possibly nested list of constraints}
}
\value{
Flat list of constraints
}
\description{
Internal function: Flatten nested lists produced by constraints_batch
}
\keyword{internal}
