% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{merge_violations}
\alias{merge_violations}
\title{Merge Violation Reports from Multiple Blocks}
\usage{
merge_violations(violation_list)
}
\arguments{
\item{violation_list}{List of violation vectors from each block}
}
\value{
Merged violation vector (sum of all blocks)
}
\description{
Internal function: Merge violation reports from multiple blocks
}
\keyword{internal}
