% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{print_constraint_summary}
\alias{print_constraint_summary}
\title{Print Constraint Summary}
\usage{
print_constraint_summary(constraints)
}
\arguments{
\item{constraints}{List of constraints}
}
\description{
Display constraint configuration summary
}
