% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{print_feedback_concise}
\alias{print_feedback_concise}
\title{Print Feedback (Concise Version)}
\usage{
print_feedback_concise(result)
}
\arguments{
\item{result}{The result object for a single list}
}
\description{
Internal function: Print concise list generation feedback
}
\keyword{internal}
